+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 <!-- version list -->
 
+## v1.16.6 (2026-08-21)
+
+### Bug Fixes
+
+- Fix calculating host python path on windows
+  ([#213](https://github.com/cloudflare/workers-py/pull/213),
+  [`47fbc48`](https://github.com/cloudflare/workers-py/commit/47fbc480e20c3fe838a286308835dd20bcf696c7))
+
+
+## v1.16.5 (2026-08-19)
+
+### Bug Fixes
+
+- Fix calculating host python path on windows
+  ([#214](https://github.com/cloudflare/workers-py/pull/214),
+  [`5e62574`](https://github.com/cloudflare/workers-py/commit/5e6257444a65d9404aaa77cd9316524fd3551df4))
+
+
 ## v1.16.4 (2026-08-13)
 
 ### Bug Fixes